// apps/region-pages/src/types/Region.ts
export type Region = {
  id: string;
  name: string;
  description?: string;
  slug: string;
  website?: string;
  email?: string;
  facebook?: string;
  twitter?: string;
  instagram?: string;
  image?: string;
  city?: string;
  state?: string;
  zip?: string;
  country?: string;
  latitude?: number;
  longitude?: number;
  zoom?: number;
};